+++ Realm/RLMRealm.mm
@@ -245,11 +245,11 @@ static void RLMRealmSetSchemaAndAlign(RLMRealm *realm, RLMSchema *targetSchema) 
                                                 File::PermissionDenied(err.UTF8String, ex.path())), outError);
                 break;
             }
-            case RealmFileException::Kind::Exists:
-                RLMSetErrorOrThrow(RLMMakeError(RLMErrorFileExists, ex), outError);
+            case RealmFileException::Kind::NotFound:
+                RLMSetErrorOrThrow(RLMMakeError(RLMErrorFileNotFound, ex), outError);
                 break;
             case RealmFileException::Kind::AccessError:
-                RLMSetErrorOrThrow(RLMMakeError(RLMErrorFileAccessError, ex), outError);
+                RLMSetErrorOrThrow(RLMMakeError(RLMErrorFileAccess, ex), outError);
                 break;
             case RealmFileException::Kind::FormatUpgradeRequired:
                 RLMSetErrorOrThrow(RLMMakeError(RLMErrorFileFormatUpgradeRequired, ex), outError);
@@ -701,9 +701,14 @@ static void CheckReadWrite(RLMRealm *realm, NSString *msg=@"Cannot write to a re
             *error = RLMMakeError(RLMErrorFileExists, ex);
         }
     }
+    catch (File::NotFound &ex) {
+        if (error) {
+            *error = RLMMakeError(RLMErrorFileNotFound, ex);
+        }
+    }
     catch (File::AccessError &ex) {
         if (error) {
-            *error = RLMMakeError(RLMErrorFileAccessError, ex);
+            *error = RLMMakeError(RLMErrorFileAccess, ex);
         }
     }
     catch (std::exception &ex) {